Make Socket a ConnectablePublisher
This doesn’t clean up as much as I had hoped.

I basically already did the work to delay connecting because I wanted to be able to setup a subscription before connecting. The reason for wanting that is that the connection might happen so fast that I miss the first message.

Either way, this is the way they like to do it and we do get a nice autoconnect() method for free, so why not.

/cc @atdrendel